pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "cna-sim"
version = "0.1.0"
description = "description"
readme = "README.md"
requires-python = ">=3.7"
license = { file = "LICENSE" }

authors = [
    { name = "wzy1935", email = "wangzengyi1935@163.com" }
]

dependencies = [
    "simpy >= 4.1.1",
    "numpy >= 2.1.3",
    "pyyaml >= 6.0"
]

[project.optional-dependencies]
influxdb = [
    "influxdb-client >= 1.48.0"
]

[project.scripts]
cna-sim = "cna_sim.cli:run"


#[project.urls]
#Homepage = "https://github.com/yourusername/mypackage"
#Documentation = "https://mypackage.readthedocs.io"
#Repository = "https://github.com/yourusername/mypackage"

